requirements.txt
flask
torch
diffusers
transformers
pillow
gunicorn
maturin